Design: 캘린더 UI 수정
